==> Scripts/Step01_ACSPull.Rmd <==
---
title: "Social Mobility: Pre-Processing"
subtitle: "Step 01: ACS Data Pull"
author: "Michael Lopez"
date: "`r format(Sys.Date(), '%B, %d, %Y'`"
output: 
  github_document: default
    toc: true
    number_sections: true
    toc_float: true
    theme: pagedown::architect
    code_folding: show
---

Load necessary libraries to import ACS census data with the tidycensus package.
```{r setup, include = FALSE, echo = FALSE}
rm(list=ls(all=TRUE))
#knitr::opts_knit$set(root.dir = 'C:/Users/micha/Documents/GitHub')
knitr::opts_knit$set(root.dir = 'C:/Users/Michael Lopez/Documents/GitHub')
#knitr::opts_knit$set(root.dir = 'C:/Users/OwenWilson-Chavez/CPAL Dropbox')
getwd()
library(tidycensus)
library(tidyverse)
library(rio)
library(tigris)
```


```{r directory test}
getwd()
```

Import necessary geocode files and transform data columns to merge with ACS dataset later in process. 
```{r geographies of interest}
acs_b <- load_variables(2019, "acs5", cache = TRUE)
acs_s <- load_variables(2014, "acs5/subject", cache = TRUE)

#fips <- tidycensus::fips_codes

years <- c(2014, 2019)

county_names <- c("Dallas County", 
                  "Rockwall County", 
                  "Collin County", 
                  "Denton County", 
                  "Tarrant County", 
                  "Kaufman County", 
                  "Ellis County", 
                  "Johnson County")

#counties <- fips %>%
#  filter(state == "TX") %>%
#  filter(county %in% county_names)

zip <- import("Social-Mobility/Data/Reference Geographies/socmob_zcta.csv")
places <- import("Social-Mobility/Data/Reference Geographies/socmob_place.csv")

colnames(zip)
colnames(places)

zip <- zip %>%
  mutate(GEOID = as.character(GEOID10)) %>%
  select(GEOID)

places <- places %>%
  mutate(GEOID = as.character(GEOID)) %>%
  select(GEOID)
```

Create dataframe of all variables to pull from census api
Bug fixed with recent tidycensus update allowing S and B table variables to pull at the same time.
```{r tidycensus variables}
socmob.old <- c(
  tot_pop = "B01003_001", #total population
  popm_18t24 = "B15001_003", #population 18 to 24 male
  popf_18t24 = "B15001_044", #population 18 to 24 female
  pop_3t4 = "B09001_004", #population between 3 to 4
  med_inc = "B19013_001", #median household income
  gini = "B19083_001", #gini coefficient
  less_hs = "B23006_002", #total population less than high school but above 9th
  hs_deg = "B23006_009", #total with high school degree
  ba_deg = "B23006_023", #total with bachelors degree or higher
  his_pop = "B03002_012", #hispanic population
  wh_pop = "B03002_003", #white population
  bl_pop = "B03002_004", #black population
  as_pop = "B03002_006", #asian population
  rohh = "B25106_024", #renter occupied households
  thh = "B25106_001", #total households
  hsm_18t24 = "B15001_006", #high school degrees between 18 and 24 male
  hsf_18t24 = "B15001_047", #high school degrees between 18 and 24 female
  es_3t4 = "B14003_004", #enrolled in school 3 to 4
  emp = "S0802_C01_001", #employed over 16
  emp_pubtr = "S0802_C04_001", #employed over 16 and take public transportation to work
  pop_bp = "S1701_C02_001", #population below poverty
  bp_u18 = "S1701_C02_002", #population under 18 below poverty
  pop_o25 = "S1501_C01_006", #population 25 and over
  pop_u18 = "S1701_C01_002", #population under 18
  emp_r25t29 = "S2301_C03_004", #employed population ratio 25 to 29
  emp_r30t34 = "S2301_C03_005", #employed population ratio 30 to 34
  emp_r35t44 = "S2301_C03_006", #employed population ratio 35 to 44
  emp_r45t54 = "S2301_C03_007" #employed population ratio 45 to 54
)

socmob.new <- c(
  tot_pop = "B01003_001", #total population
  popm_18t24 = "B15001_003", #population 18 to 24 male
  popf_18t24 = "B15001_044", #population 18 to 24 female
  pop_3t4 = "B09001_004", #population between 3 to 4
  med_inc = "B19013_001", #median household income
  gini = "B19083_001", #gini coefficient
  less_hs = "B23006_002", #total population less than high school but above 9th
  hs_deg = "B23006_009", #total with high school degree
  ba_deg = "B23006_023", #total with bachelors degree or higher
  his_pop = "B03002_012", #hispanic population
  wh_pop = "B03002_003", #white population
  bl_pop = "B03002_004", #black population
  as_pop = "B03002_006", #asian population
  rohh = "B25106_024", #renter occupied households
  thh = "B25106_001", #total households
  hsm_18t24 = "B15001_006", #high school degrees between 18 and 24 male
  hsf_18t24 = "B15001_047", #high school degrees between 18 and 24 female
  es_3t4 = "B14003_004", #enrolled in school 3 to 4
  emp = "S0802_C01_001", #employed over 16
  emp_pubtr = "S0802_C04_001", #employed over 16 and take public transportation to work
  pop_bp = "S1701_C02_001", #population below poverty
  bp_u18 = "S1701_C02_002", #population under 18 below poverty
  pop_o25 = "S1501_C01_006", #population 25 and over
  pop_u18 = "S0101_C01_022", #population under 18
  emp_r25t29 = "S2301_C03_004", #employed population ratio 25 to 29
  emp_r30t34 = "S2301_C03_005", #employed population ratio 30 to 34
  emp_r35t44 = "S2301_C03_006", #employed population ratio 35 to 44
  emp_r45t54 = "S2301_C03_007" #employed population ratio 45 to 54
)
```

Tidycensus pull of all zip codes of interest.
```{r zip pull}
socmob.zip.2014 <- get_acs(
    geography = "zcta", 
    variables = socmob.old,
    year = 2014, 
    survey = "acs5", 
    output = "wide") %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_14"))) %>%
  rename(NAME = NAME_14) %>%
  mutate(TYPE = "ZIP")

socmob.zip.2019 <- get_acs(
    geography = "zcta", 
    variables = socmob.new,
    year = 2019, 
    survey = "acs5", 
    output = "wide") %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_19"))) %>%
  rename(NAME = NAME_19) %>%
  mutate(TYPE = "ZIP")

socmob.zip <- socmob.zip.2014 %>%
  select(-GEOID) %>%
  left_join(., socmob.zip.2019)

rm(socmob.zip.2014)
rm(socmob.zip.2019)

socmob.zip <- inner_join(zip, socmob.zip)

rm(zip)
```

Tidycensus pull of all place geographies of interest.
```{r place pull}
socmob.place.2014 <- get_acs(
    geography = "place", 
    variables = socmob.old,
    year = 2014, 
    survey = "acs5", 
    output = "wide") %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_14"))) %>%
  rename(NAME = NAME_14) %>%
  mutate(TYPE = "PLACE")

socmob.place.2019 <- get_acs(
    geography = "place", 
    variables = socmob.new,
    year = 2019, 
    survey = "acs5", 
    output = "wide") %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_19"))) %>%
  rename(NAME = NAME_19) %>%
  mutate(TYPE = "PLACE")

socmob.place <- full_join(socmob.place.2014, socmob.place.2019)

rm(socmob.place.2014)
rm(socmob.place.2019)

socmob.place <- inner_join(places, socmob.place)

rm(places)
```

Tidycensus pull of all county geographies of interest.
```{r county pull}
socmob.counties.2014 <- get_acs(
    geography = "county", 
    state = "TX",
    county = county_names,
    variables = socmob.old,
    year = 2014, 
    survey = "acs5", 
    output = "wide") %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_14"))) %>%
  rename(NAME = NAME_14) %>%
  mutate(TYPE = "COUNTY")

socmob.counties.2019 <- get_acs(
    geography = "county", 
    state = "TX",
    county = county_names,
    variables = socmob.new,
    year = 2019, 
    survey = "acs5", 
    output = "wide") %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_19"))) %>%
  rename(NAME = NAME_19) %>%
  mutate(TYPE = "COUNTY")

socmob.counties <- full_join(socmob.counties.2014, socmob.counties.2019)

rm(socmob.counties.2014)
rm(socmob.counties.2019)
```

Tidycensus pull of all census tracts of interest.
Not able to easily pull multiple census blocks with standard pull and instead it's necessary to run a loop with `purrr::map_dfr` for all counties of interest. Then run code chunk for each year of interest before merging multiple groups.
```{r tract pull}
socmob.tract.2014 <- map_dfr(
  .x = county_names,
  ~get_acs(
    geography = "tract", 
    year = 2014,
    state = "TX",
    county = .x,
    variables = socmob.old,
    survey = "acs5", 
    output = "wide")
  )

socmob.tract.2019 <- map_dfr(
  .x = county_names,
  ~get_acs(
    geography = "tract", 
    year = 2019,
    state = "TX",
    county = .x,
    variables = socmob.new,
    survey = "acs5", 
    output = "wide")
  )

socmob.tract.2014 <- socmob.tract.2014 %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_14"))) %>%
  rename(NAME = NAME_14) %>%
  mutate(TYPE = "TRACT")

socmob.tract.2019 <- socmob.tract.2019 %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_19"))) %>%
  rename(NAME = NAME_19) %>%
  mutate(TYPE = "TRACT")

socmob.tract <- full_join(socmob.tract.2014, socmob.tract.2019)

rm(socmob.tract.2014)
rm(socmob.tract.2019)
```

Combine all ACS geographies into one file in order to calculate percent change variables and percentage variables.
Once new variables have been calculated split dataframe back into groups for each geography type.
```{r variable calculations}
socmob.combine <- full_join(full_join(full_join(socmob.zip, socmob.place), socmob.tract), socmob.counties)

names(socmob.combine)
socmob.calc <- socmob.combine %>%
  mutate(pop_18t24E_14 = popm_18t24E_14+popf_18t24E_14,
         pop_18t24E_19 = popm_18t24E_19+popf_18t24E_19,
         wh_per_19 = wh_popE_19/tot_popE_19,
         bl_per_19 = bl_popE_19/tot_popE_19,
         as_per_19 = as_popE_19/tot_popE_19,
         his_per_19 = his_popE_19/tot_popE_19,
         wh_per_14 = wh_popE_14/tot_popE_14,
         bl_per_14 = bl_popE_14/tot_popE_14,
         as_per_14 = as_popE_14/tot_popE_14,
         his_per_14 = his_popE_14/tot_popE_14,
         adj_med_inc_14 = med_incE_14*(1.08),                    #Inflation rate 2019/2014
         emp_r_14 = ((emp_r25t29E_14+emp_r30t34E_14+emp_r35t44E_14+emp_r45t54E_14)/4)/100,
         emp_r_19 = ((emp_r25t29E_19+emp_r30t34E_19+emp_r35t44E_19+emp_r45t54E_19)/4)/100,
         hs_18t24_per_14 = (hsm_18t24E_14+hsf_18t24E_14)/pop_18t24E_14,
         hs_18t24_per_19 = (hsm_18t24E_19+hsm_18t24E_19)/pop_18t24E_19,
         es_3t4_per_14 = es_3t4E_14/pop_3t4E_14,
         es_3t4_per_19 = es_3t4E_19/pop_3t4E_19,
         emp_pubtr_per_14 = emp_pubtrE_14/empE_14,
         emp_pubtr_per_19 = emp_pubtrE_19/empE_19,
         bp_per_14 = pop_bpE_14/tot_popE_14,
         bp_per_19 = pop_bpE_19/tot_popE_19,
         cbp_per_14 = bp_u18E_14/pop_u18E_14,
         cbp_per_19 = bp_u18E_19/pop_u18E_19,
         rohh_per_14 = rohhE_14/thhE_14,
         rohh_per_19 = rohhE_19/thhE_19,
         per_rohh = ((rohh_per_19 - rohh_per_14)/rohh_per_14),
         per_tot_pop = ((tot_popE_19 - tot_popE_14)/tot_popE_14),
         per_med_inc = ((med_incE_19 - adj_med_inc_14)/adj_med_inc_14),
         per_gini = ((giniE_19 - giniE_14)/giniE_14),
         per_less_hs = ((less_hsE_19 - less_hsE_14)/less_hsE_14),
         per_hs_deg = ((hs_degE_19 - hs_degE_14)/hs_degE_14),
         per_wh = ((wh_popE_19 - wh_popE_14)/wh_popE_14),
         per_bl = ((bl_popE_19 - bl_popE_14)/bl_popE_14),
         per_as = ((as_popE_19 - as_popE_14)/as_popE_14),
         per_his = ((his_popE_19 - his_popE_14)/his_popE_14),
         per_emp_r = emp_r_19-emp_r_14,
         per_hs_18t24 = (((hsm_18t24E_19+hsm_18t24E_19)-(hsm_18t24E_14+hsf_18t24E_14))/(hsm_18t24E_14+hsf_18t24E_14)),
         per_es_3t4 = ((es_3t4E_19 - es_3t4E_14)/es_3t4E_14),
         per_emp_pubtr = ((emp_pubtrE_19 - emp_pubtrE_14)/emp_pubtrE_14),
         per_bp = ((pop_bpE_19 - pop_bpE_14)/pop_bpE_14),
         per_cbp = ((bp_u18E_19 - bp_u18E_14)/bp_u18E_14)
         ) %>%
  mutate_if(is.numeric, funs(replace_na(., 0))) %>%
  select(-(ends_with("_14")), -(ends_with("M_19")), -(emp_r25t29E_19:emp_r45t54E_19), -(popm_18t24E_19:popf_18t24E_19), -(hsm_18t24E_19:hsf_18t24E_19))

perchange <- socmob.combine %>%
  group_by(TYPE) %>%
  summarise_if(is.numeric, sum) %>%
  mutate(pop_18t24E_14 = popm_18t24E_14+popf_18t24E_14,
         pop_18t24E_19 = popm_18t24E_19+popf_18t24E_19,
         wh_per_19 = wh_popE_19/tot_popE_19,
         bl_per_19 = bl_popE_19/tot_popE_19,
         as_per_19 = as_popE_19/tot_popE_19,
         his_per_19 = his_popE_19/tot_popE_19,
         wh_per_14 = wh_popE_14/tot_popE_14,
         bl_per_14 = bl_popE_14/tot_popE_14,
         as_per_14 = as_popE_14/tot_popE_14,
         his_per_14 = his_popE_14/tot_popE_14,
         adj_med_inc_14 = med_incE_14*(1.08),                    #Inflation rate 2019/2014
         emp_r_14 = ((emp_r25t29E_14+emp_r30t34E_14+emp_r35t44E_14+emp_r45t54E_14)/4)/100,
         emp_r_19 = ((emp_r25t29E_19+emp_r30t34E_19+emp_r35t44E_19+emp_r45t54E_19)/4)/100,
         hs_18t24_per_14 = (hsm_18t24E_14+hsf_18t24E_14)/pop_18t24E_14,
         hs_18t24_per_19 = (hsm_18t24E_19+hsm_18t24E_19)/pop_18t24E_19,
         es_3t4_per_14 = es_3t4E_14/pop_3t4E_14,
         es_3t4_per_19 = es_3t4E_19/pop_3t4E_19,
         emp_pubtr_per_14 = emp_pubtrE_14/empE_14,
         emp_pubtr_per_19 = emp_pubtrE_19/empE_19,
         bp_per_14 = pop_bpE_14/tot_popE_14,
         bp_per_19 = pop_bpE_19/tot_popE_19,
         cbp_per_14 = bp_u18E_14/pop_u18E_14,
         cbp_per_19 = bp_u18E_19/pop_u18E_19,
         rohh_per_14 = rohhE_14/thhE_14,
         rohh_per_19 = rohhE_19/thhE_19,
         per_rohh = ((rohh_per_19 - rohh_per_14)/rohh_per_14),
         per_tot_pop = ((tot_popE_19 - tot_popE_14)/tot_popE_14),
         per_med_inc = ((med_incE_19 - adj_med_inc_14)/adj_med_inc_14),
         per_gini = ((giniE_19 - giniE_14)/giniE_14),
         per_less_hs = ((less_hsE_19 - less_hsE_14)/less_hsE_14),
         per_hs_deg = ((hs_degE_19 - hs_degE_14)/hs_degE_14),
         per_wh = ((wh_popE_19 - wh_popE_14)/wh_popE_14),
         per_bl = ((bl_popE_19 - bl_popE_14)/bl_popE_14),
         per_as = ((as_popE_19 - as_popE_14)/as_popE_14),
         per_his = ((his_popE_19 - his_popE_14)/his_popE_14),
         #per_emp_r = emp_r_19-emp_r_14,
         per_hs_18t24 = (((hsm_18t24E_19+hsm_18t24E_19)-(hsm_18t24E_14+hsf_18t24E_14))/(hsm_18t24E_14+hsf_18t24E_14)),
         per_es_3t4 = ((es_3t4E_19 - es_3t4E_14)/es_3t4E_14),
         per_emp_pubtr = ((emp_pubtrE_19 - emp_pubtrE_14)/emp_pubtrE_14),
         per_bp = ((pop_bpE_19 - pop_bpE_14)/pop_bpE_14),
         per_cbp = ((bp_u18E_19 - bp_u18E_14)/bp_u18E_14)
         ) %>%
  select(TYPE, starts_with("per_")) %>%
  export("C:/Users/Michael Lopez/Documents/GitHub/Social-Mobility/Data/Exploration/PercentChange_AggregatedbyType.csv")

names(socmob.calc)

socmob.zip <- socmob.calc %>%
  filter(str_detect(TYPE, "ZIP"))

socmob.place <- socmob.calc %>%
  filter(str_detect(TYPE, "PLACE"))

socmob.counties <- socmob.calc %>%
  filter(str_detect(TYPE, "COUNTY"))

socmob.tract <- socmob.calc %>%
  filter(str_detect(TYPE, "TRACT"))

rm(socmob.combine)
```

Export to csv
```{r}
setwd("C:/Users/Michael Lopez/Documents/GitHub")

export(socmob.zip, "Social-Mobility/Data/ACS/SocialMobility_Zip.csv")
export(socmob.place, "Social-Mobility/Data/ACS/SocialMobility_Place.csv")
export(socmob.tract, "Social-Mobility/Data/ACS/SocialMobility_Tract.csv")
export(socmob.counties, "Social-Mobility/Data/ACS/SocialMobility_County.csv")
```
