==> Scripts/Step02_DatatogeoJSON.Rmd <==
---
title: "Social Mobility: Pre-Processing"
subtitle: "Step 02: ACS Data to geoJSON"
author: "Michael Lopez"
date: "`r format(Sys.Date(), '%B, %d, %Y'`"
output: 
  github_document: default
    toc: true
    number_sections: true
    toc_float: true
    theme: pagedown::architect
    code_folding: show
---

```{r setup, include = FALSE, echo = FALSE}
rm(list=ls(all=TRUE))
knitr::opts_knit$set(root.dir = 'C:/Users/Michael Lopez/Documents/GitHub')
#knitr::opts_knit$set(root.dir = 'C:/Users/OwenWilson-Chavez/CPAL Dropbox')
getwd()
library(tidyverse)
library(rio)
library(sf)
```

```{r directory test}
getwd()
```

```{r import data}
#import csv files of data
socmob.zip.csv <- import("Social-Mobility/Data/ACS/SocialMobility_Zip.csv")
socmob.place.csv <- import("Social-Mobility/Data/ACS/SocialMobility_Place.csv")
socmob.tract.csv <- import("Social-Mobility/Data/ACS/SocialMobility_Tract.csv")
socmob.county.csv <- import("Social-Mobility/Data/ACS/SocialMobility_County.csv")

#import shape files
zip.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/sm_zcta.shp")
place.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/sm_places.shp")
tract.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/sm_tracts.shp")
counties.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/sm_counties.shp")
school.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/sm_schooldistricts.shp")
statehouse.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/Texas_House_of_Representative_Districts.shp")
statesenate.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/Texas_State_Senate_Districts.shp")
congressdistrict.shape <- st_read("Social-Mobility/Data/Reference Geographies/Shapefiles/Texas_Congressional_Districts.shp")

plot(school.shape$geometry)

```

```{r transform and merge}
#convert shape GEOID columns from factor to numeric to help merge
zip.shape <- zip.shape %>%
  mutate(GEOID = as.numeric(GEOID10))

place.shape <- place.shape %>%
  mutate(GEOID = as.numeric(GEOID))

tract.shape <- tract.shape %>%
  mutate(GEOID = as.numeric(GEOID))

counties.shape <- counties.shape %>%
  mutate(GEOID = as.numeric(GEOID))

socmob.zip.csv <- socmob.zip.csv %>%
  mutate(GEOID = as.numeric(GEOID))

socmob.place.csv <- socmob.place.csv %>%
  mutate(GEOID = as.numeric(GEOID))

socmob.tract.csv <- socmob.tract.csv %>%
  mutate(GEOID = as.numeric(GEOID))

socmob.county.csv <- socmob.county.csv %>%
  mutate(GEOID = as.integer(GEOID))


#merge csv and shape files into shape file
zip.merge <- left_join(zip.shape, socmob.zip.csv, by = "GEOID")

place.merge <- left_join(place.shape, socmob.place.csv, by = "GEOID")

tract.merge <- left_join(tract.shape, socmob.tract.csv, by = "GEOID")

county.merge <- left_join(counties.shape, socmob.county.csv, by = "GEOID")

#delete unnecessary columns
names(zip.merge)
zip.select <- zip.merge %>%
  select(fid, GEOID, NAME, tot_popE_19:geometry) %>%
  mutate(label = GEOID)

names(place.merge)
place.select <- place.merge %>%
  select(fid, GEOID, NAME.x, tot_popE_19:geometry) %>%
  rename(NAME = NAME.x) %>%
  mutate(label = GEOID)

names(tract.merge)
tract.select <- tract.merge %>%
  select(fid, GEOID, NAME.x, tot_popE_19:geometry) %>%
  rename(NAME = NAME.x) %>%
  mutate(label = GEOID)

names(county.merge)
county.select <- county.merge %>%
  select(fid, GEOID, NAME.x, tot_popE_19:geometry) %>%
  rename(NAME = NAME.x) %>%
  mutate(label = GEOID)

school.merge <- school.shape %>%
  .[counties.shape, ] %>%
  select(OBJECTID, NAME, geometry) %>%
  rename(fid = OBJECTID) %>%
  mutate(label = NAME)

statehouse.merge <- statehouse.shape %>%
  .[counties.shape, ] %>%
  rename(fid = OBJECTID,
         NAME = District) %>%
  select(fid, NAME, geometry) %>%
  mutate(NAME = paste("State House District", NAME),
         label = NAME)

statesenate.merge <- statesenate.shape %>%
  .[counties.shape, ] %>%
  rename(fid = OBJECTID,
         NAME = District) %>%
  select(fid, NAME, geometry) %>%
  mutate(NAME = paste("State Senate District", NAME),
         label = NAME)

congressdistrict.merge <- congressdistrict.shape %>%
  .[counties.shape, ] %>%
  rename(fid = OBJECTID,
         NAME = District) %>%
  select(fid, NAME, geometry) %>%
  mutate(NAME = paste("Congressional District", NAME),
         label = NAME)
```


```{r geojson export}
getwd()
geojsonio::geojson_write(school.merge, file = "Social-Mobility/Data/geojson/sm_schooldistricts.geojson")
school.merge %>%
  st_point_on_surface(.) %>%
  geojsonio::geojson_write(., file = "Social-Mobility/Data/geojson/schooldistricts_pt.geojson")




geojsonio::geojson_write(zip.merge, file = "Social-Mobility/Data/Intermediate/sm_zcta.geojson")
zip.merge %>%
  st_point_on_surface(.) %>%
  geojsonio::geojson_write(., file = "Social-Mobility/Data/Intermediate/zcta_pt.geojson")




geojsonio::geojson_write(place.merge, file = "Social-Mobility/Data/Intermediate/sm_places.geojson")
place.merge %>%
  st_point_on_surface(.) %>%
geojsonio::geojson_write(., file = "Social-Mobility/Data/Intermediate/places_pt.geojson")




geojsonio::geojson_write(tract.merge, file = "Social-Mobility/Data/Intermediate/sm_tracts.geojson")
tract.merge %>%
  st_point_on_surface(.) %>%
geojsonio::geojson_write(., file = "Social-Mobility/Data/Intermediate/tracts_pt.geojson")




geojsonio::geojson_write(county.merge, file = "Social-Mobility/Data/Intermediate/sm_counties.geojson")
county.merge %>%
  st_point_on_surface(.) %>%
geojsonio::geojson_write(., file = "Social-Mobility/Data/Intermediate/counties_pt.geojson")




geojsonio::geojson_write(statehouse.merge, file = "Social-Mobility/Data/geojson/sm_statehouse.geojson")
statehouse.merge %>%
  st_point_on_surface(.) %>%
geojsonio::geojson_write(., file = "Social-Mobility/Data/geojson/statehouse_pt.geojson")




geojsonio::geojson_write(statesenate.merge, file = "Social-Mobility/Data/geojson/sm_statesenate.geojson")
statesenate.merge %>%
  st_point_on_surface(.) %>%
geojsonio::geojson_write(., file = "Social-Mobility/Data/geojson/statesenate_pt.geojson")




geojsonio::geojson_write(congressdistrict.merge, file = "Social-Mobility/Data/geojson/sm_fedcongress.geojson")
congressdistrict.merge %>%
  st_point_on_surface(.) %>%
geojsonio::geojson_write(., file = "Social-Mobility/Data/geojson/fedcongress_pt.geojson")
```
