==> Scripts/Step03_FeaturestogeoJSON.Rmd <==
---
title: "Social Mobility: Pre-Processing"
subtitle: "Step 03: Feature Points to geoJSON"
author: "Michael Lopez"
date: "`r format(Sys.Date(), '%B, %d, %Y'`"
output: 
  github_document: default
    toc: true
    number_sections: true
    toc_float: true
    theme: pagedown::architect
    code_folding: show
---

```{r setup, include = FALSE, echo = FALSE}
rm(list=ls(all=TRUE))
#knitr::opts_knit$set(root.dir = 'E:/CPAL Dropbox')
knitr::opts_knit$set(root.dir = 'C:/Users/micha/CPAL Dropbox')
#knitr::opts_knit$set(root.dir = 'C:/Users/OwenWilson-Chavez/CPAL Dropbox')
getwd()
library(tidyverse)
library(rio)
library(tigris)
library(sf)
library(lubridate)
```

```{r directory test}
getwd()
```

```{r, echo = FALSE}
#import geography comparison shape file
zip.sh <- st_read("C:/Users/micha/Documents/GitHub/Social-Mobility/Data/Reference Geographies/Shapefiles/sm_zcta.shp")

county.sh <- st_read("C:/Users/micha/Documents/GitHub/Social-Mobility/Data/Reference Geographies/Shapefiles/sm_counties.shp")

#import feature shape files
aftersch.sh <- st_read("Data Library/Dallas Afterschool/DallasAfterschool_Programs_Dec2019.shp")
recctr.sh <- st_read("Data Library/04_CPAL Features/Places/Places.gdb", layer = "CommRecCtrs")
snap.sh <- st_read("Data Library/USDA/GIS/SNAP_Store_Locations_TX_Dec2020.shp")
banks.sh <- st_read("Data Library/Texas Department of Banking/Geocoded/NorthTexas_BankingLocations_Feb2021.shp") 
comcen.sh <- st_read("Data Library/NCTCOG/Archive/CommunityCenters.shp")
comhealth.sh <- st_read("Data Library/Miscellaneous/Community Health Clinics/CommunityHealthClincs_June2020.shp")
creditunions.sh <- st_read("Data Library/Google/GIS/Shapefiles/CreditUnions_Dec2020.shp") 
supermarket.sh <- st_read("Data Library/Miscellaneous/Grocery Stores/FoodandGrocery_DallasCounty_July2020.shp")
highered.sh <- st_read("Data Library/NCTCOG/Archive/HigherEducation.shp")
pharmacies.sh <-st_read("Data Library/Texas State Board of Pharmacy/GIS/TBSP_NTX_Dec2020.shp")
libraries.sh <- st_read("Data Library/City of Dallas/02_Boundaries and Features/Libraries.shp")
librariesTX.sh <- st_read("Data Library/Texas Public Libraries/Texas_Public_Libraries.shp")
schools.sh <- st_read("Data Library/Texas Education Agency/GIS/Current_Schools_2020-2021.shp")
schoolsprivate.sh <- st_read("Data Library/Texas Private School Accreditation Commission/TexasPrivateSchools_Feb2021.shp")
wic.sh <- st_read("Data Library/Google/GIS/Shapefiles/WIC_Locations_Dec2020.shp") 
childcare.sh <- st_read("Data Library/Texas Department of Family and Protective Services/ChildcareCenters_SubsidizedTX_Feb2021.shp")
childcareAll.sh <- st_read("Data Library/Texas Department of Family and Protective Services/ChildcareCenters_unSubsidizedTX_Feb2021.shp")
childcareAll.sh <- as.data.frame(childcareAll.sh) %>%
  select(-geometry) %>%
  mutate(x_coor = X,
         y_coor = Y) %>%
  st_as_sf(x = ., coords = c(x = "X", y = "Y"),
                        crs = "+proj=longlat +datum=WGS84 +no_defs")
vita.sh <- st_read("Data Library/IRS/VITA Centers Feb 2021/VITA_Centers_NTX.shp")
fqhc.sh <- st_read("Data Library/Texas Department of State Health Services/Federally Qualified Health Center/February 2021/FederallyQualifiedHealthCenters_NTX.shp")

```

```{r}
#edit shape files to only contain necessary columns
snap.join <- snap.sh %>%
  mutate(Display = "SNAP Retailer",
         variable = "snap") %>%
  rename(Name = Store_Name) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(snap.sh)

names(banks.sh)

banks.join <- banks.sh %>%
  st_transform(crs = 4326) %>%
  mutate(Display = "Bank",
         variable = "bank") %>%
  rename(Longitude = X,
         Latitude = Y,
         Name = USER_entit,
         Address = USER_addre) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(banks.sh)

comcen.join <- comcen.sh %>%
  st_transform(crs = 4326) %>%
  mutate(Longitude = st_coordinates(.)[,1],
         Latitude = st_coordinates(.)[,2],
         Display = "Community Center",
         variable = "comcen") %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(comcen.sh)

comhealth.join <- comhealth.sh %>%
  mutate(Display = "Community Health Clinic",
         variable = "comhel") %>%
  rename(Name = USER_Clini,
         Address = StAddr,
         Longitude = X,
         Latitude = Y) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(comhealth.sh)

creditunions.join <- creditunions.sh %>%
  mutate(Longitude = st_coordinates(.)[,1],
         Latitude = st_coordinates(.)[,2],
         Display = "Credit Union",
         City = word(plcs_dd,-1),
         variable = "crunion") %>%
  rename(Name = plcs_nm,
         Address = plcs_dd) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
st_crs(creditunions.join) = 4326
rm(creditunions.sh)

supermarket.join <- supermarket.sh %>%
  mutate(Display = "Supermarket",
         variable = "market") %>%
  rename(Name = BusinessNa,
         Address = Bus_Addres) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(supermarket.sh)

pharmacies.join <- pharmacies.sh %>%
  mutate(Display = "Pharmacy",
         variable = "pharm") %>%
  rename(Address = StAddr,
         Longitude = X,
         Latitude = Y,
         Name = USER_phy_n) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(pharmacies.sh)

librariesTX.join <- librariesTX.sh %>%
  select(USER_Libra, USER_Addre, X, Y, USER_City) %>%
  mutate(Display = "Library",
         variable = "lib") %>%
  rename(Name = USER_Libra,
         Address = USER_Addre,
         City = USER_City,
         Longitude = X,
         Latitude = Y) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(librariesTX.sh)

libraries.join <- libraries.sh %>%
  mutate(Display = "Library",
         City = "Dallas",
         variable = "lib") %>%
  rename(Name = LIBRARY,
         Address = ADDRESS,
         Longitude = X,
         Latitude = Y) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry) %>%
  full_join(., as.data.frame(librariesTX.join))
rm(libraries.sh)
rm(librariesTX.join)

schools.join <- schools.sh %>%
  rename(Address = StAddr,
         Name = School_Nam,
         Longitude = X,
         Latitude = Y,
         District_Type = District_T,
         Grade_Level = Grade_Leve
         ) %>%
  mutate(Display = ifelse(District_Type == "INDEPENDENT" & Grade_Level == "High School", "Independent High School",
                          ifelse(District_Type == "INDEPENDENT" & Grade_Level %in% c("Middle", "Junior High"), "Independent Middle School",
                                 ifelse(District_Type == "INDEPENDENT" & Grade_Level %in% c("Elementary", "Elementary/Secondary"), "Independent Elementary School",
                                        ifelse(District_Type == "CHARTER" & Grade_Level == "High School", "Charter High School",
                                               ifelse(District_Type == "CHARTER" & Grade_Level %in% c("Middle", "Junior High"), "Charter Middle School",
                                                      ifelse(District_Type == "CHARTER" & Grade_Level %in% c("Elementary", "Elementary/Secondary"), "Charter Elementary School",
                                                             "Other School")))))),
         variable = ifelse(Display == "Independent High School", "ihs",
                          ifelse(Display == "Independent Middle School", "ims",
                                 ifelse(Display == "Independent Elementary School", "ies",
                                        ifelse(Display == "Charter High School", "chs",
                                               ifelse(Display == "Charter Middle School", "cms",
                                                      ifelse(Display == "Charter Elementary School", "ces",
                                                             "os"))))))) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry) %>%
  filter(variable != "os")
rm(schools.sh)

schoolsprivate.join <- schoolsprivate.sh %>%
  mutate(Display = "Private School",
         variable = "prs",
         Address = ShortLabel) %>%
  rename(Longitude = X,
         Latitude = Y,
         Name = USER_Schoo) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)

rm(schoolsprivate.sh)

wic.join <- wic.sh %>%
  mutate(Longitude = st_coordinates(.)[,1],
         Latitude = st_coordinates(.)[,2],
         Display = "WIC Clinic",
         City = word(plcs_dd,-1),
         variable = "wic") %>%
  rename(Name = plcs_nm,
         Address = plcs_dd) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
st_crs(wic.join) = 4326
rm(wic.sh)

names(childcare.sh)

childcare.join <- childcare.sh %>%
  mutate(Display = "Subsidized Childcare Center",
         variable = "scc") %>%
  rename(Longitude = X,
         Latitude = Y,
         Name = USER_Opera,
         Address = USER_Addre) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(childcare.sh)

childcareAll.join <- childcareAll.sh %>%
  mutate(Display = "Childcare Center",
         variable = "uscc") %>%
  rename(Longitude = x_coor,
         Latitude = y_coor,
         Name = USER_Opera,
         Address = USER_Addre) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(childcareAll.sh)

vita.join <- vita.sh %>%
  mutate(Name = USER_Name,
         Address = StAddr,
         Display = "Volunteer Income Tax Assistance Center",
         variable = "vita") %>%
  rename(Longitude = X,
         Latitude = Y) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(vita.sh)

fqhc.join <- fqhc.sh %>%
  mutate(Name = USER_name,
         Address = StAddr,
         Display = "Federally Qualified Health Center",
         variable = "fqhc") %>%
  rename(Longitude = X,
         Latitude = Y) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
rm(fqhc.sh)

highered.join <- highered.sh %>%
  mutate(Display = "College Campuses",
         variable = "highed",
         Longitude = st_coordinates(.)[,1],
         Latitude = st_coordinates(.)[,2]) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry)
st_crs(highered.join, crs = 4326)
rm(highered.sh)

aftersch.join <- aftersch.sh %>%
  st_transform(crs = 4326) %>%
  mutate(Display = "Afterschool Program",
         variable = "aftersch",
         Address = NA,
         City = "Dallas") %>%
  rename(Longitude = longitude,
         Latitude = latitude,
         Name = Account_Na) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry) %>%
  st_transform(crs = 4326)
rm(aftersch.sh)

recctr.join <- recctr.sh %>%
  mutate(Display = "Recreation Center",
         variable = "recctr",
         Longitude = st_coordinates(.)[,1],
         Latitude = st_coordinates(.)[,2]) %>%
  rename(Address = ADDRESS,
         Name = NAME,
         geometry = Shape) %>%
  select(Display, variable, Name, Address, City, Longitude, Latitude, geometry) %>%
  st_transform(crs = 4326)
rm(recctr.sh)

```

```{r}
#select points within geography of interest
recctr.county <- recctr.join[county.sh, ]
rm(recctr.join)

aftersch.county <- aftersch.join[county.sh, ]
rm(aftersch.join)

highered.county <- highered.join[county.sh, ]
rm(highered.join)

snap.county <- snap.join[county.sh, ]
rm(snap.join)

banks.county <- banks.join[county.sh, ]
rm(banks.join)

comcen.county <- comcen.join[county.sh, ]
rm(comcen.join)

comhealth.county <- comhealth.join[county.sh, ]
rm(comhealth.join)

creditunions.county <- creditunions.join[county.sh, ]
rm(creditunions.join)

supermarket.county <- supermarket.join[county.sh, ]
rm(supermarket.join)

pharmacies.county <- pharmacies.join[county.sh, ]
rm(pharmacies.join)

libraries.county <- libraries.join[county.sh, ]
rm(libraries.join)

schools.county <- schools.join[county.sh, ]
rm(schools.join)

schoolsprivate.county <- schoolsprivate.join[county.sh, ]
rm(schoolsprivate.join)

wic.county <- wic.join[county.sh, ]
rm(wic.join)

childcare.county <- childcare.join[county.sh, ]
rm(childcare.join)

childcareAll.county <- childcareAll.join[county.sh, ]
rm(childcareAll.join)

vita.county <- vita.join[county.sh, ]
rm(vita.join)

fqhc.county <- fqhc.join[county.sh, ]
rm(fqhc.join)
```

```{r}
#join all files into one
featureset <- rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(rbind(
                    banks.county, comcen.county), comhealth.county), creditunions.county), supermarket.county), pharmacies.county), libraries.county), schools.county), schoolsprivate.county), wic.county), childcare.county), childcareAll.county), vita.county), fqhc.county), highered.county), recctr.county), aftersch.county)

featureset <- featureset %>%
  mutate(FID = 1:n(),
         GEOID = FID,
         Name = str_to_title(Name),
         Latitude = round(Latitude, digits = 4),
         Longitude = round(Longitude, digits = 4)) %>%
  as.data.frame(.) %>%
  select(FID, everything(), -geometry) %>%
  st_as_sf(., coords = c("Longitude", "Latitude"), crs = 4326)
```


```{r}
#export to geojson
geojsonio::geojson_write(featureset, file = "C:/Users/micha/Documents/GitHub/Social-Mobility/Data/geojson/sm_featureset.geojson")
```
