==> Scripts/Step04_DataCalculation.Rmd <==
---
title: "Social Mobility: Pre-Processing"
subtitle: "Step 02: ACS Data Transformation"
author: "Michael Lopez"
date: "`r format(Sys.Date(), '%B, %d, %Y'`"
output: 
  github_document: default
    toc: true
    number_sections: true
    toc_float: true
    theme: pagedown::architect
    code_folding: show
---

Setup of script, load necessary libraries and set working directory.
```{r setup, include = FALSE, echo = FALSE}
rm(list=ls(all=TRUE))
#knitr::opts_knit$set(root.dir = 'C:/Users/micha/Documents/GitHub')
knitr::opts_knit$set(root.dir = 'C:/Users/Michael Lopez/Documents/GitHub')
#knitr::opts_knit$set(root.dir = 'C:/Users/OwenWilson-Chavez/CPAL Dropbox')
getwd()
library(tidyverse)
library(rio)
library(sf)
```

Doublecheck and make certain that working directory has changed.
```{r directory test}
getwd()
```

Import four different geography types and rename columns to join with field types in other data types. Remove duplicate data columns as well.
```{r import geography data, echo=FALSE}
#import acs data
sm_zcta <- st_read("Social-Mobility/Data/Intermediate/sm_zcta.geojson")
sm_places <- st_read("Social-Mobility/Data/Intermediate/sm_places.geojson")
sm_tracts <- st_read("Social-Mobility/Data/Intermediate/sm_tracts.geojson")
sm_counties <- st_read("Social-Mobility/Data/Intermediate/sm_counties.geojson")

#delete unnecessary columns
names(sm_zcta)
zip.select <- sm_zcta %>%
  select(fid, NAME, GEOID10, ALAND10, AWATER10, INTPTLAT10, INTPTLON10, tot_popE_19:geometry) %>%
  rename(GEOID = GEOID10,
         ALAND = ALAND10,
         AWATER = AWATER10,
         INTPTLAT = INTPTLAT10,
         INTPTLON = INTPTLON10) %>%
  mutate(GEOID = as.integer(GEOID))

names(sm_places)
place.select <- sm_places %>%
  select(fid, GEOID, NAME.x, ALAND:INTPTLON, tot_popE_19:geometry) %>%
  rename(NAME = NAME.x)

names(sm_tracts)
tract.select <- sm_tracts %>%
  select(fid, GEOID, NAME.x, ALAND:INTPTLON, tot_popE_19:geometry) %>%
  rename(NAME = NAME.x)

names(sm_counties)
county.select <- sm_counties %>%
  select(fid, GEOID, NAME.x, ALAND:INTPTLON, tot_popE_19:geometry) %>%
  rename(NAME = NAME.x)

rm(sm_zcta)
rm(sm_places)
rm(sm_tracts)
rm(sm_counties)
```

Change working directory for the following code chunk in order to import data from a different filepath on CPAL dropbox. All imports of new data should go in this block chunk.
```{r import raw data, echo=FALSE}
#setwd("C:/Users/micha/CPAL Dropbox") #Michael
setwd("E:/CPAL Dropbox") #Michael
#setwd("C:/Users/OwenWilson-Chavez/CPAL Dropbox") #Owen

#Urban Institute Rental Assistance Priority Index
renter <- import("Data Library/Urban Institute/Rental Assistance Priority Index/housing_index_state_adj.csv")

#HUD Housing Choice Vouchers
housing_vouchers <- st_read("Data Library/HUD/Housing Choice Vouchers/Housing_Choice_Vouchers_by_Tract-shp/HCV_BY_TRACT.shp")

#County Voter Precincts
precincts_collin <- st_read("Data Library/Texas Elections/Collin County Elections/2018_CollinCounty_VoterPrecincts.geojson")
precincts_dallas <- st_read("Data Library/Texas Elections/Dallas County Elections/Precinct Shapefiles/Precinct13.shp")
precincts_denton <- st_read("Data Library/Texas Elections/Denton County Elections/Precinct Shapefiles/Voter_Precincts_-_2018.shp")
#precincts_ellis
precincts_johnson <- st_read("Data Library/Texas Elections/Johnson County Elections/Precinct Shapefiles/Voting_Precincts_of_Johnson_County.shp")
#precincts_kaurman
precincts_rockwall <- st_read("Data Library/Texas Elections/Rockwall County Elections/Precinct Shapefiles/2012_Precinct_Boundaries_Rockwall_County.shp")
precincts_tarrant <- st_read("Data Library/Texas Elections/Tarrant County Elections/Precinct Shapefiles/Tarrant_County_Elections_Voting_Precinct_Boundary.shp")

#County Voter Data
vote2016 <- import("Data Library/Texas Elections/Dallas County Elections/Votes by Precinct/2016_Vote_byPrecinct_DallasCounty.csv")
#vote2018 <- import("Data Library/Dallas County Elections/Votes by Precinct/2018_Vote_byPrecinct_DallasCounty.csv")
vote2020 <- import("Data Library/Texas Elections/Dallas County Elections/Votes by Precinct/2020_Vote_byPrecinct_DallasCounty.csv")

#Brookings EITC Data
eitc_zip <- import("Basic Needs/04_Projects/EITC/Brookings EITC/zipcodeTX_TY14.csv")
eitc_county <- import("Basic Needs/04_Projects/EITC/Brookings EITC/countyTX_TY14.csv")
eitc_place <- import("Basic Needs/04_Projects/EITC/Brookings EITC/cityTX_TY14.csv")
```

```{r housing vouchers}
names(housing_vouchers)
housing_lim <- as.data.frame(housing_vouchers) %>%
  select(GEOID, HCV_PUBLIC) %>%
  mutate(GEOID = as.numeric(GEOID))
  
names(tract.select)
tract_housing <- tract.select %>%
  left_join(., housing_lim) %>%
  mutate(HCV_rohh = HCV_PUBLIC/rohhE_19)

rm(housing_vouchers)
rm(housing_lim)
```

```{r renter instability}
names(renter)

renter_lim <- renter %>%
  select(GEOID, housing_index) %>%
  mutate(GEOID = as.numeric(GEOID),
         housing_index = round(housing_index, digits = 4)) %>%
  rename(rent_instability = housing_index)

tract_rent <- left_join(tract.select, renter_lim)

rm(renter_lim)
rm(renter)
```

```{r voter data Dallas County}
names(precincts_dallas)
precincts_dallas <- precincts_dallas %>%
  select(FID_Precin, PCT, geometry) %>%
  rename(precinct = PCT,
         FID = FID_Precin)

vote2016 <- vote2016 %>%
  mutate(precinct = str_sub(precinct, 1, 4)) %>%
  select(precinct, total_ballots, registered_voters) %>%
  group_by(precinct) %>%
  summarise(total_ballots_2016 = sum(total_ballots),
            registered_voters_2016 = sum(registered_voters),
            turnout_2016 = total_ballots_2016/registered_voters_2016) %>%
  drop_na() %>%
  filter(!is.infinite(turnout_2016))

vote2020 <- vote2020 %>%
  mutate(precinct = str_sub(precinct, 1, 4)) %>%
  select(precinct, total_ballots, registered_voters) %>%
  group_by(precinct) %>%
  summarise(total_ballots_2020 = sum(total_ballots),
            registered_voters_2020 = sum(registered_voters),
            turnout_2020 = total_ballots_2020/registered_voters_2020) %>%
  drop_na() %>%
  filter(!is.infinite(turnout_2020))

vote_change <- vote2020 %>%
  full_join(., vote2016) %>%
  mutate(turnout_dif = turnout_2020-turnout_2016) %>%
  left_join(precincts_dallas, .)
```

```{r brookings EITC}
eitc_zip <- eitc_zip %>%
  mutate(GEOID = as.numeric(str_sub(ZIPCODE, 4, 8))) %>%
  select(GEOID, ERETURN14, EEIC14, EEICAM14, ECTC14, ECTCAM14) %>%
  rename(Tot_Returns = ERETURN14,
         Tot_EITC = EEIC14,
         Sum_EITC = EEICAM14,
         Tot_CTC = ECTC14,
         Sum_CTC = ECTCAM14) %>%
  mutate(Tot_Returns = as.numeric(gsub(",", "", Tot_Returns)),
         Tot_EITC = as.numeric(gsub(",", "", Tot_EITC)),
         Sum_EITC = as.numeric(gsub(",", "", Sum_EITC)),
         Tot_CTC = as.numeric(gsub(",", "", Tot_CTC)),
         Sum_CTC = as.numeric(gsub(",", "", Sum_CTC)),
         Avg_EITC = hablar::rationalize(Sum_EITC/Tot_EITC),
         Avg_CTC = hablar::rationalize(Sum_CTC/Tot_CTC)) %>%
left_join(zip.select, .)

eitc_place <- eitc_place %>%
  mutate(GEOID = as.numeric(str_c(STFIPS, str_sub(plc_fip, 4, 8)))) %>%
  select(GEOID, ERETURN14, EEIC14, EEICAM14, ECTC14, ECTCAM14) %>%
  rename(Tot_Returns = ERETURN14,
         Tot_EITC = EEIC14,
         Sum_EITC = EEICAM14,
         Tot_CTC = ECTC14,
         Sum_CTC = ECTCAM14) %>%
  mutate(Tot_Returns = as.numeric(gsub(",", "", Tot_Returns)),
         Tot_EITC = as.numeric(gsub(",", "", Tot_EITC)),
         Sum_EITC = as.numeric(gsub(",", "", Sum_EITC)),
         Tot_CTC = as.numeric(gsub(",", "", Tot_CTC)),
         Sum_CTC = as.numeric(gsub(",", "", Sum_CTC)),
         Avg_EITC = hablar::rationalize(Sum_EITC/Tot_EITC),
         Avg_CTC = hablar::rationalize(Sum_CTC/Tot_CTC)) %>%
left_join(place.select, .)

eitc_county <- eitc_county %>%
  mutate(GEOID = as.numeric(str_c(STFIPS, str_sub(county_fips, 4, 6)))) %>%
  select(GEOID, ERETURN14, EEIC14, EEICAM14, ECTC14, ECTCAM14) %>%
  rename(Tot_Returns = ERETURN14,
         Tot_EITC = EEIC14,
         Sum_EITC = EEICAM14,
         Tot_CTC = ECTC14,
         Sum_CTC = ECTCAM14) %>%
  mutate(Tot_Returns = as.numeric(gsub(",", "", Tot_Returns)),
         Tot_EITC = as.numeric(gsub(",", "", Tot_EITC)),
         Sum_EITC = as.numeric(gsub(",", "", Sum_EITC)),
         Tot_CTC = as.numeric(gsub(",", "", Tot_CTC)),
         Sum_CTC = as.numeric(gsub(",", "", Sum_CTC)),
         Avg_EITC = hablar::rationalize(Sum_EITC/Tot_EITC),
         Avg_CTC = hablar::rationalize(Sum_CTC/Tot_CTC)) %>%
left_join(county.select, .)
```

```{r join to each other}
names(tract_rent)

new_tract_var <- tract_rent %>%
  left_join(.,  as.data.frame(tract_housing)) %>%
  mutate(per_rohh = hablar::rationalize(as.numeric(per_rohh)),
         per_tot_pop = hablar::rationalize(as.numeric(per_tot_pop)),
         per_med_inc = hablar::rationalize(as.numeric(per_med_inc)),
         per_gini = hablar::rationalize(as.numeric(per_gini)),
         per_less_hs = hablar::rationalize(as.numeric(per_less_hs)),
         per_hs_deg = hablar::rationalize(as.numeric(per_hs_deg)),
         per_wh = hablar::rationalize(as.numeric(per_wh)),
         per_bl = hablar::rationalize(as.numeric(per_bl)),
         per_as = hablar::rationalize(as.numeric(per_as)),
         per_his = hablar::rationalize(as.numeric(per_his)),
         per_hs_18t24 = hablar::rationalize(as.numeric(per_hs_18t24)),
         per_es_3t4 = hablar::rationalize(as.numeric(per_es_3t4)),
         per_emp_pubtr = hablar::rationalize(as.numeric(per_emp_pubtr)),
         per_bp = hablar::rationalize(as.numeric(per_bp)),
         per_cbp = hablar::rationalize(as.numeric(per_cbp)),
         HCV_rohh = hablar::rationalize(as.numeric(HCV_rohh)))

new_zip_var <- zip.select %>%
  left_join(.,  as.data.frame(eitc_zip)) %>%
  mutate(per_rohh = hablar::rationalize(as.numeric(per_rohh)),
         per_tot_pop = hablar::rationalize(as.numeric(per_tot_pop)),
         per_med_inc = hablar::rationalize(as.numeric(per_med_inc)),
         per_gini = hablar::rationalize(as.numeric(per_gini)),
         per_less_hs = hablar::rationalize(as.numeric(per_less_hs)),
         per_hs_deg = hablar::rationalize(as.numeric(per_hs_deg)),
         per_wh = hablar::rationalize(as.numeric(per_wh)),
         per_bl = hablar::rationalize(as.numeric(per_bl)),
         per_as = hablar::rationalize(as.numeric(per_as)),
         per_his = hablar::rationalize(as.numeric(per_his)),
         per_hs_18t24 = hablar::rationalize(as.numeric(per_hs_18t24)),
         per_es_3t4 = hablar::rationalize(as.numeric(per_es_3t4)),
         per_emp_pubtr = hablar::rationalize(as.numeric(per_emp_pubtr)),
         per_bp = hablar::rationalize(as.numeric(per_bp)),
         per_cbp = hablar::rationalize(as.numeric(per_cbp)),)

new_place_var <- place.select  %>%
  left_join(.,  as.data.frame(eitc_place)) %>%
  mutate(per_rohh = hablar::rationalize(as.numeric(per_rohh)),
         per_tot_pop = hablar::rationalize(as.numeric(per_tot_pop)),
         per_med_inc = hablar::rationalize(as.numeric(per_med_inc)),
         per_gini = hablar::rationalize(as.numeric(per_gini)),
         per_less_hs = hablar::rationalize(as.numeric(per_less_hs)),
         per_hs_deg = hablar::rationalize(as.numeric(per_hs_deg)),
         per_wh = hablar::rationalize(as.numeric(per_wh)),
         per_bl = hablar::rationalize(as.numeric(per_bl)),
         per_as = hablar::rationalize(as.numeric(per_as)),
         per_his = hablar::rationalize(as.numeric(per_his)),
         per_hs_18t24 = hablar::rationalize(as.numeric(per_hs_18t24)),
         per_es_3t4 = hablar::rationalize(as.numeric(per_es_3t4)),
         per_emp_pubtr = hablar::rationalize(as.numeric(per_emp_pubtr)),
         per_bp = hablar::rationalize(as.numeric(per_bp)),
         per_cbp = hablar::rationalize(as.numeric(per_cbp)))

new_county_var <- county.select %>%
  left_join(.,  as.data.frame(eitc_county))
```

```{r acs modifications}
#names(new_tract_var)
#belpov <- new_tract_var %>%
#  filter(bp_per_18 >= 0.4) %>%
#  select(pop_bpE_18,geometry) %>%
#  rename(bp_o40 = pop_bpE_18)

#new_zip_var <- zip.select %>%
#  select(GEOID, tot_popE_18, geometry) %>%
#  st_intersection(., belpov) %>%
#  group_by(GEOID) %>%
#  summarise(bp_o40 = sum(bp_o40),
#            tot_popE_18 = sum(tot_popE_18)) %>%
#  mutate(bp_o40per = bp_o40/tot_popE_18,
#         GEOID = as.numeric(GEOID))

#new_place_var <- place.select %>%
#  select(GEOID, tot_popE_18, geometry) %>%
#  st_intersection(., belpov) %>%
#  group_by(GEOID) %>%
#  summarise(bp_o40 = sum(bp_o40),
#            tot_popE_18 = sum(tot_popE_18)) %>%
#  mutate(bp_o40per = bp_o40/tot_popE_18)

#new_county_var <-  county.select %>%
#  select(GEOID, tot_popE_18, geometry) %>%
#  st_intersection(., belpov) %>%
#  group_by(GEOID) %>%
#  summarise(bp_o40 = sum(bp_o40),
#            tot_popE_18 = sum(tot_popE_18)) %>%
#  mutate(bp_o40per = bp_o40/tot_popE_18)
#  
```

Function which will group each variable into bins between 0 and 4 based on the standard deviation of each variable.
Cuts occur at +/- 0.5 standard deviations and +/- 1 standard deviation.
```{r sdcut function}
sdcut <- function(x) {
  sd1p <- mean(x, na.rm = TRUE)+(sd(x, na.rm = TRUE)*0.25)
  sd1n <- mean(x, na.rm = TRUE)-(sd(x, na.rm = TRUE)*0.25)
  sd2p <- mean(x, na.rm = TRUE)+(sd(x, na.rm = TRUE)*1)
  sd2n <- mean(x, na.rm = TRUE)-(sd(x, na.rm = TRUE)*1)
  ifelse(x > sd2p, 4,
         ifelse(x > sd1p & x < sd2p, 3,
                ifelse(x > sd1n & x < sd1p, 2,
                       ifelse(x > sd2n & x < sd1n, 1, 0))))
}
```

Convert all estimate variables by `log()` in order to then apply sdcut function.
```{r sdcut all geographies}
#zip codes
names(new_zip_var)
socmob.zip.sd <- as.data.frame(new_zip_var) %>%
  mutate(giniE_19 = giniE_19*(-1),
         per_gini = per_gini*(-1)) %>%
  mutate_at(c(8:62), funs(c(log(.+10)))) %>%
  mutate(bp_u18E_19 = bp_u18E_19*(-1),
         pop_bpE_19 = pop_bpE_19*(-1),
         bp_per_19 = bp_per_19*(-1),
         cbp_per_19 = cbp_per_19*(-1),
         per_bp = per_bp*(-1),
         per_cbp = per_cbp*(-1)) %>%
  mutate_at(c(8:62), funs(c(sdcut(.)))) %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_sd"))) %>%
  rename(NAME = NAME_sd,
         GEOID = GEOID_sd,
         ALAND = ALAND_sd,
         AWATER = AWATER_sd,
         INTPTLAT = INTPTLAT_sd,
         INTPTLON = INTPTLON_sd) %>%
  select(-geometry_sd)
socmob.zip.upload <- left_join(new_zip_var, socmob.zip.sd)

#places
socmob.place.sd <- as.data.frame(new_place_var) %>%
  mutate(giniE_19 = giniE_19*(-1),
         per_gini = per_gini*(-1)) %>%
  mutate_at(c(8:62), funs(c(log(.+10)))) %>%
  mutate(bp_u18E_19 = bp_u18E_19*(-1),
         pop_bpE_19 = pop_bpE_19*(-1),
         bp_per_19 = bp_per_19*(-1),
         cbp_per_19 = cbp_per_19*(-1),
         per_bp = per_bp*(-1),
         per_cbp = per_cbp*(-1)) %>%
  mutate_at(c(8:62), funs(c(sdcut(.)))) %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_sd"))) %>%
  rename(NAME = NAME_sd,
         GEOID = GEOID_sd,
         ALAND = ALAND_sd,
         AWATER = AWATER_sd,
         INTPTLAT = INTPTLAT_sd,
         INTPTLON = INTPTLON_sd) %>%
  select(-geometry_sd)

socmob.place.upload <- left_join(new_place_var, socmob.place.sd)

#tracts
socmob.tract.sd <- as.data.frame(new_tract_var) %>%
  mutate(rent_instability = rent_instability*(-1),
         giniE_19 = giniE_19*(-1),
         per_gini = per_gini*(-1)) %>%
  mutate_at(c(8:58), funs(c(log(.+10)))) %>%
  mutate(bp_u18E_19 = bp_u18E_19*(-1),
         pop_bpE_19 = pop_bpE_19*(-1),
         bp_per_19 = bp_per_19*(-1),
         cbp_per_19 = cbp_per_19*(-1),
         per_bp = per_bp*(-1),
         per_cbp = per_cbp*(-1)) %>%
  mutate_at(c(8:58), funs(c(sdcut(.)))) %>%
  setNames(c(names(.)[1], paste0(names(.)[-1],"_sd"))) %>%
  rename(NAME = NAME_sd,
         GEOID = GEOID_sd,
         ALAND = ALAND_sd,
         AWATER = AWATER_sd,
         INTPTLAT = INTPTLAT_sd,
         INTPTLON = INTPTLON_sd) %>%
  select(-geometry_sd)

socmob.tract.upload <- left_join(new_tract_var, socmob.tract.sd)
```

```{r descriptive stats}
socmob.calc <- full_join(full_join(full_join(as.data.frame(socmob.tract.upload), as.data.frame(socmob.zip.upload)), as.data.frame(socmob.place.upload)),as.data.frame(new_county_var))

str(new_county_var$GEOID)
str(socmob.zip.upload$GEOID)
str(socmob.tract.upload$GEOID)
str(socmob.place.upload$GEOID)

var <-  c("nbr.val", "nbr.null", "nbr.na", "min", "max", "range", "sum", "median", "mean", "SE.mean", "CI.mean.0.95", "var", "std.dev", "coef.var")
socmob.calc.d <- cbind(var, pastecs::stat.desc(socmob.calc))

n <- socmob.calc.d$var

# transpose all but the first column (name)
socmob.calc.d <- as.data.frame(t(socmob.calc.d[,-1]))
colnames(socmob.calc.d) <- n
socmob.calc.d$variable <- factor(row.names(socmob.calc.d))
socmob.calc.d <- socmob.calc.d %>%
  select(variable, min, max, range, mean, median, std.dev)
setwd("C:/Users/Michael Lopez/Documents/GitHub")
datadic <- import("Social-Mobility/DataDictionary.csv")
datadic_join <- datadic %>%
  select(variable:highisgood) %>%
  full_join(., socmob.calc.d)
```

```{r geojson export}
setwd("C:/Users/Michael Lopez/Documents/GitHub")
export(datadic_join, "Social-Mobility/DataDictionary.csv")
geojsonio::geojson_write(socmob.zip.upload, file = "Social-Mobility/Data/geojson/sm_zcta.geojson")
geojsonio::geojson_write(socmob.place.upload, file = "Social-Mobility/Data/geojson/sm_places.geojson")
geojsonio::geojson_write(socmob.tract.upload, file = "Social-Mobility/Data/geojson/sm_tracts.geojson")
geojsonio::geojson_write(new_county_var, file = "Social-Mobility/Data/geojson/sm_counties.geojson")
```
