==> Scripts/Step06_TimeSeriesPull.Rmd <==
---
title: "NTX Child Poverty Over Time by Tract"
subtitle: "UNTD"
author: "Michael Lopez"
date: "`r format(Sys.Date(), '%B, %d, %Y'`"
output: 
  github_document: default
    toc: true
    number_sections: true
    toc_float: true
    theme: pagedown::architect
    code_folding: show
---

Load necessary libraries to import ACS census data with the tidycensus package.

```{r setup, include = FALSE, echo = FALSE}
rm(list=ls(all=TRUE))
knitr::opts_knit$set(root.dir = 'C:/Users/Michael Lopez/Documents/GitHub/Social-Mobility/')
#knitr::opts_knit$set(root.dir = 'C:/Users/micha/Documents/GitHub/Social-Mobility/')
getwd()
library(tidycensus)
library(tidyverse)
library(rio)
library(sf)
library(ggthemes)
library(tigris)
library(directlabels)
```

```{r directory test, include = FALSE, echo = FALSE}
getwd()
```

```{r setup for graphs, include = FALSE, echo = FALSE}
acs_b <- load_variables(2015, "acs5", cache = TRUE)
acs_s <- load_variables(2019, "acs5/subject", cache = TRUE)
#view(acs18)
#data(fips_codes)
years_old <- lst(2012, 2013, 2014, 2015, 2016)
years_new <- lst(2017, 2018, 2019)

counties <- c("Dallas", "Tarrant", "Denton", "Collin", "Rockwall", "Kaufman", "Ellis", "Johnson")

###############################################################################################
###############################################################################################

CPAL.colors = c("#008097", "#ec008c", "#eaca2d", "#b4b4b4", "#9bd9e3", "#fdddd7")
CPAL.colors.Rev = c("#fdddd7", "#9bd9e3", "#b4b4b4", "#eaca2d", "#ec008c", "#008097")
Black = c("#000000", "#000000", "#000000")

###############################################################################################
###############################################################################################
theme_cpal <- function(base_size = 12, base_family = "sans") {
  colors <- deframe(ggthemes::ggthemes_data[["fivethirtyeight"]])
  (theme_foundation(base_size = base_size, base_family = base_family)
    + theme(
      line = element_line(colour = "#b4b4b4"),
      rect = element_rect(fill = "#ffffff",
                          linetype = 1, colour = NA),
      text = element_text(family = "Roboto", face = "bold", colour = "#6c6c6c"),
      axis.title = element_text(),
      axis.title.x = element_text(vjust = 2),
      axis.title.y = element_text(vjust = 2),
      axis.text = element_text(color = "#b4b4b4"),
      axis.ticks = element_blank(),
      #axis.ticks.length = unit(6, "pt"),
      axis.line = element_line(color = "#b4b4b4", size = 1.5, linetype = "solid"),
      legend.background = element_rect(),
      legend.position = "none",
      legend.direction = "horizontal",
      legend.box = "horizontal",
      panel.grid.major = element_line(colour = "#e1e1e1"),
      panel.grid.minor = element_blank(),
      plot.title = element_text(hjust = 0, size = rel(1.5), face = "bold"),
      plot.margin = unit(c(1, 6, 1, 1), "lines"),
      panel.border = element_rect(size=1, fill = NA),
      strip.background = element_rect()
    ))
}
###############################################################################################
###############################################################################################
```

```{r tidycensus variable selection, include = FALSE, echo = FALSE}
acs_new <- c(
  tot_pop = "B01003_001", #total population
  popm_18t24 = "B15001_003", #population 18 to 24 male
  popf_18t24 = "B15001_044", #population 18 to 24 female
  pop_3t4 = "B09001_004", #population between 3 to 4
  med_inc = "B19013_001", #median household income
  gini = "B19083_001", #gini coefficient
  less_hs = "B23006_002", #total population less than high school but above 9th
  hs_deg = "B23006_009", #total with high school degree
  ba_deg = "B23006_023", #total with bachelors degree or higher
  his_pop = "B03002_012", #hispanic population
  wh_pop = "B03002_003", #white population
  bl_pop = "B03002_004", #black population
  as_pop = "B03002_006", #asian population
  rohh = "B25106_024", #renter occupied households
  thh = "B25106_001", #total households
  hsm_18t24 = "B15001_006", #high school degrees between 18 and 24 male
  hsf_18t24 = "B15001_047", #high school degrees between 18 and 24 female
  es_3t4 = "B14003_004", #enrolled in school 3 to 4
  emp = "S0802_C01_001", #employed over 16
  emp_pubtr = "S0802_C04_001", #employed over 16 and take public transportation to work
  pop_bp = "S1701_C02_001", #population below poverty
  bp_u18 = "S1701_C02_002", #population under 18 below poverty
  pop_o25 = "S1501_C01_006", #population 25 and over
  pop_u18 = "S0101_C01_022", #population under 18
  emp_r25t29 = "S2301_C03_004", #employed population ratio 25 to 29
  emp_r30t34 = "S2301_C03_005", #employed population ratio 30 to 34
  emp_r35t44 = "S2301_C03_006", #employed population ratio 35 to 44
  emp_r45t54 = "S2301_C03_007" #employed population ratio 45 to 54
  )

acs_old <- c(
  tot_pop = "B01003_001", #total population
  popm_18t24 = "B15001_003", #population 18 to 24 male
  popf_18t24 = "B15001_044", #population 18 to 24 female
  pop_3t4 = "B09001_004", #population between 3 to 4
  med_inc = "B19013_001", #median household income
  gini = "B19083_001", #gini coefficient
  less_hs = "B23006_002", #total population less than high school but above 9th
  hs_deg = "B23006_009", #total with high school degree
  ba_deg = "B23006_023", #total with bachelors degree or higher
  his_pop = "B03002_012", #hispanic population
  wh_pop = "B03002_003", #white population
  bl_pop = "B03002_004", #black population
  as_pop = "B03002_006", #asian population
  rohh = "B25106_024", #renter occupied households
  thh = "B25106_001", #total households
  hsm_18t24 = "B15001_006", #high school degrees between 18 and 24 male
  hsf_18t24 = "B15001_047", #high school degrees between 18 and 24 female
  es_3t4 = "B14003_004", #enrolled in school 3 to 4
  emp = "S0802_C01_001", #employed over 16
  emp_pubtr = "S0802_C04_001", #employed over 16 and take public transportation to work
  pop_bp = "S1701_C02_001", #population below poverty
  bp_u18 = "S1701_C02_002", #population under 18 below poverty
  pop_o25 = "S1501_C01_006", #population 25 and over
  pop_u18 = "S1701_C01_002", #population under 18
  emp_r25t29 = "S2301_C03_004", #employed population ratio 25 to 29
  emp_r30t34 = "S2301_C03_005", #employed population ratio 30 to 34
  emp_r35t44 = "S2301_C03_006", #employed population ratio 35 to 44
  emp_r45t54 = "S2301_C03_007" #employed population ratio 45 to 54
  )
```

```{r pull variables from tidycensus, include = FALSE, echo = FALSE}
cpal_tract_old <- map(
  years_old,
  ~get_acs(
    geography = "tract",
    state = "TX",
    #county = counties,
    variables = acs_old,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_old, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_tract_new <- map(
  years_new,
  ~get_acs(
    geography = "tract",
    state = "TX",
    #county = counties,
    variables = acs_new,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_new, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_tract <- rbind(cpal_tract_old, cpal_tract_new)
rm(cpal_tract_new)
rm(cpal_tract_old)

untd_tract_long <- cpal_tract %>%
  filter(str_detect(NAME, str_c(counties, collapse = "|"))) %>%
  mutate(pop_18t24E = popm_18t24E+popf_18t24E,
         wh_per = wh_popE/tot_popE,
         bl_per = bl_popE/tot_popE,
         as_per = as_popE/tot_popE,
         his_per = his_popE/tot_popE,
         emp_r = ((emp_r25t29E+emp_r30t34E+emp_r35t44E+emp_r45t54E)/4)/100,
         hs_18t24_per = (hsm_18t24E+hsm_18t24E)/pop_18t24E,
         es_3t4_per = es_3t4E/pop_3t4E,
         emp_pubtr_per = emp_pubtrE/empE,
         bp_per = pop_bpE/tot_popE,
         cbp_per = bp_u18E/pop_u18E,
         rohh_per = rohhE/thhE,
         type = "Tract") %>%
  select(-ends_with("M"), -(emp_r25t29E:emp_r45t54M))

untd_tract_wide <- untd_tract_long %>%
  pivot_wider(names_from = "year", 
              values_from = c(tot_popE, med_incE, giniE, pop_u18E, pop_bpE, bp_u18E, bp_per, cbp_per, rohh_per, emp_pubtr_per, es_3t4_per, hs_18t24_per, emp_r, his_per, as_per, bl_per, wh_per, pop_18t24E))
```

```{r pull variables from tidycensus, include = FALSE, echo = FALSE}
zip <- import("Data/Reference Geographies/socmob_zcta.csv") %>%
  select(ZIP) %>%
  rename(GEOID = ZIP) %>%
  mutate(GEOID = as.character(GEOID))

cpal_zip_old <- map(
  years_old,
  ~get_acs(
    geography = "zcta",
    variables = acs_old,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_old, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_zip_new <- map(
  years_new,
  ~get_acs(
    geography = "zcta",
    variables = acs_new,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_new, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_zip <- cpal_zip_old %>%
  mutate(GEOID = as.character(as.numeric(GEOID)-4800000)) %>%
  rbind(., cpal_zip_new) %>%
  inner_join(., zip)

rm(cpal_zip_new)
rm(cpal_zip_old)

untd_zip_long <- cpal_zip %>%
  mutate(pop_18t24E = popm_18t24E+popf_18t24E,
         wh_per = wh_popE/tot_popE,
         bl_per = bl_popE/tot_popE,
         as_per = as_popE/tot_popE,
         his_per = his_popE/tot_popE,
         emp_r = ((emp_r25t29E+emp_r30t34E+emp_r35t44E+emp_r45t54E)/4)/100,
         hs_18t24_per = (hsm_18t24E+hsm_18t24E)/pop_18t24E,
         es_3t4_per = es_3t4E/pop_3t4E,
         emp_pubtr_per = emp_pubtrE/empE,
         bp_per = pop_bpE/tot_popE,
         cbp_per = bp_u18E/pop_u18E,
         rohh_per = rohhE/thhE,
         type = "Zip") %>%
  select(-ends_with("M"))

untd_zip_wide <- untd_zip_long %>%
  pivot_wider(names_from = "year", 
              values_from = c(tot_popE, med_incE, giniE, pop_u18E, pop_bpE, bp_u18E, bp_per, cbp_per, rohh_per, emp_pubtr_per, es_3t4_per, hs_18t24_per, emp_r, his_per, as_per, bl_per, wh_per, pop_18t24E))
```

```{r pull variables from tidycensus, include = FALSE, echo = FALSE}
place <- import("Data/Reference Geographies/socmob_place.csv") %>%
  select(GEOID) %>%
  mutate(GEOID = as.character(GEOID))

cpal_place_old <- map(
  years_old,
  ~get_acs(
    geography = "place",
    state = "TX",
    variables = acs_old,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_old, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_place_new <- map(
  years_new,
  ~get_acs(
    geography = "place",
    state = "TX",
    variables = acs_new,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_new, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_place <- rbind(cpal_place_old, cpal_place_new) %>%
  inner_join(place, .)

rm(cpal_place_new)
rm(cpal_place_old)

untd_place_long <- cpal_place %>%
  mutate(pop_18t24E = popm_18t24E+popf_18t24E,
         wh_per = wh_popE/tot_popE,
         bl_per = bl_popE/tot_popE,
         as_per = as_popE/tot_popE,
         his_per = his_popE/tot_popE,
         emp_r = ((emp_r25t29E+emp_r30t34E+emp_r35t44E+emp_r45t54E)/4)/100,
         hs_18t24_per = (hsm_18t24E+hsm_18t24E)/pop_18t24E,
         es_3t4_per = es_3t4E/pop_3t4E,
         emp_pubtr_per = emp_pubtrE/empE,
         bp_per = pop_bpE/tot_popE,
         cbp_per = bp_u18E/pop_u18E,
         rohh_per = rohhE/thhE,
         type = "Place") %>%
  select(-ends_with("M"))
names(untd_place_long)

untd_place_wide <- untd_place_long %>%
  pivot_wider(names_from = "year", 
              values_from = c(tot_popE, med_incE, giniE, pop_u18E, pop_bpE, bp_u18E, bp_per, cbp_per, rohh_per, emp_pubtr_per, es_3t4_per, hs_18t24_per, emp_r, his_per, as_per, bl_per, wh_per, pop_18t24E))
```

```{r pull variables from tidycensus, include = FALSE, echo = FALSE}
county_names <- c("Dallas County", 
                  "Rockwall County", 
                  "Collin County", 
                  "Denton County", 
                  "Tarrant County", 
                  "Kaufman County", 
                  "Ellis County", 
                  "Johnson County")

cpal_county_old <- map(
  years_old,
  ~get_acs(
    geography = "county",
    state = "TX",
    county = county_names,
    variables = acs_old,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_old, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_county_new <- map(
  years_new,
  ~get_acs(
    geography = "county",
    state = "TX",
    county = county_names,
    variables = acs_new,
    year = .x, 
    survey = "acs5", 
    output = "wide"),
) %>%
  map2(years_new, ~mutate(.x, year = .y)) %>%
  reduce(., rbind)

cpal_county <- rbind(cpal_county_old, cpal_county_new)

rm(cpal_county_new)
rm(cpal_county_old)

untd_county_long <- cpal_county %>%
  mutate(pop_18t24E = popm_18t24E+popf_18t24E,
         wh_per = wh_popE/tot_popE,
         bl_per = bl_popE/tot_popE,
         as_per = as_popE/tot_popE,
         his_per = his_popE/tot_popE,
         emp_r = ((emp_r25t29E+emp_r30t34E+emp_r35t44E+emp_r45t54E)/4)/100,
         hs_18t24_per = (hsm_18t24E+hsm_18t24E)/pop_18t24E,
         es_3t4_per = es_3t4E/pop_3t4E,
         emp_pubtr_per = emp_pubtrE/empE,
         bp_per = pop_bpE/tot_popE,
         cbp_per = bp_u18E/pop_u18E,
         rohh_per = rohhE/thhE,
         type = "County") %>%
  select(-ends_with("M"))

untd_county_wide <- untd_county_long %>%
  pivot_wider(names_from = "year", 
              values_from = c(tot_popE, med_incE, giniE, pop_u18E, pop_bpE, bp_u18E, bp_per, cbp_per, rohh_per, emp_pubtr_per, es_3t4_per, hs_18t24_per, emp_r, his_per, as_per, bl_per, wh_per, pop_18t24E))
```

```{r}
untd_wide <- full_join(full_join(full_join(untd_tract_wide, untd_zip_wide), untd_place_wide), untd_county_wide)

untd_long <- full_join(full_join(full_join(untd_tract_long, untd_zip_long), untd_place_long), untd_county_long)
```


```{r}
export(untd_tract_wide, "Data/ACS/TimeSeriesbyTract_Wide.csv")
export(untd_tract_long, "Data/ACS/TimeSeriesbyTract_Long.csv")

export(untd_zip_wide, "Data/ACS/TimeSeriesbyZip_Wide.csv")
export(untd_zip_long, "Data/ACS/TimeSeriesbyZip_Long.csv")

export(untd_place_wide, "Data/ACS/TimeSeriesbyPlace_Wide.csv")
export(untd_place_long, "Data/ACS/TimeSeriesbyPlace_Long.csv")

export(untd_county_wide, "Data/ACS/TimeSeriesbyCounty_Wide.csv")
export(untd_county_long, "Data/ACS/TimeSeriesbyCounty_Long.csv")

export(untd_wide, "Data/ACS/TimeSeriesAllGeographies_Wide.csv")
export(untd_long, "Data/ACS/TimeSeriesAllGeographies_Long.csv")
```
